fix compile issues and some warnings with clang
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@3873 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ dialogs/Properties/PropertyDialog.h
@@ -223,7 +223,7 @@ public:
         combo.GetLBText(combo.GetCurSel(), buf);
         for (size_t i = 0; i < contentList.size(); i++)
             if (contentList[i].compare(buf)==0)
-                return i;
+                return (int)i;
         return -1;
     }
